Use doubles instead of floats in class BrownianMotionPose

+++ src/hand-tracking/include/BrownianMotionPose.h
@@ -10,9 +10,9 @@
 class BrownianMotionPose : public bfl::StateModel
 {
 public:
-    BrownianMotionPose(const float q_xy, const float q_z, const float theta, const float cone_angle, const unsigned int seed) noexcept;
+    BrownianMotionPose(const double q_xy, const double q_z, const double theta, const double cone_angle, const unsigned int seed) noexcept;
 
-    BrownianMotionPose(const float q_xy, const float q_z, const float theta, const float cone_angle) noexcept;
+    BrownianMotionPose(const double q_xy, const double q_z, const double theta, const double cone_angle) noexcept;
 
     BrownianMotionPose() noexcept;
 
@@ -37,22 +37,22 @@ public:
     bool setProperty(const std::string& property) override { return false; };
 
 protected:
-    float                                 q_xy_;        /* Noise standard deviation for z   3D position */
-    float                                 q_z_;         /* Noise standard deviation for x-y 3D position */
-    float                                 theta_;       /* Noise standard deviation for axis-angle rotation */
-    float                                 cone_angle_;  /* Noise standard deviation for axis-angle axis cone */
+    double                                 q_xy_;        /* Noise standard deviation for z   3D position */
+    double                                 q_z_;         /* Noise standard deviation for x-y 3D position */
+    double                                 theta_;       /* Noise standard deviation for axis-angle rotation */
+    double                                 cone_angle_;  /* Noise standard deviation for axis-angle axis cone */
 
     Eigen::Vector4f                       cone_dir_;    /* Cone direction of rotation. Fixed, left here for future implementation. */
 
     std::mt19937_64                       generator_;
-    std::normal_distribution<float>       distribution_pos_xy_;
-    std::normal_distribution<float>       distribution_pos_z_;
-    std::normal_distribution<float>       distribution_theta_;
-    std::uniform_real_distribution<float> distribution_cone_;
-    std::function<float()>                gaussian_random_pos_xy_;
-    std::function<float()>                gaussian_random_pos_z_;
-    std::function<float()>                gaussian_random_theta_;
-    std::function<float()>                gaussian_random_cone_;
+    std::normal_distribution<double>       distribution_pos_xy_;
+    std::normal_distribution<double>       distribution_pos_z_;
+    std::normal_distribution<double>       distribution_theta_;
+    std::uniform_real_distribution<double> distribution_cone_;
+    std::function<double()>                gaussian_random_pos_xy_;
+    std::function<double()>                gaussian_random_pos_z_;
+    std::function<double()>                gaussian_random_theta_;
+    std::function<double()>                gaussian_random_cone_;
 
     void addAxisangleDisturbance(const Eigen::Ref<const Eigen::MatrixXd>& disturbance_vec, Eigen::Ref<Eigen::MatrixXd> current_vec);
 };
